Fix launcher web release wasm compatibility
