+++ .pm/tasks/task_1d90a16ad56f4c4e96bfa5370199b3d7.execution.md
@@ -1,0 +1,23 @@
+# task_1d90a16ad56f4c4e96bfa5370199b3d7 Execution Log
+
+- task_uid: task_1d90a16ad56f4c4e96bfa5370199b3d7
+- title: Fix release build-web-dist launcher wasm compatibility
+- owner_role: viewer_engineer
+- worktree_hint: /home/scc/worktrees/oasis7-testing-release-web-dist-launcher-wasm-compat
+
+<!-- Append entries using:
+Example:
+  ## YYYY-MM-DD HH:MM:SS CST / role_name
+  - 完成内容: ...
+  - 遗留事项: ...
+-->
+## 2026-04-13 23:39:30 CST / viewer_engineer
+- 完成内容: 复盘 `Release Packages` run `24349229764`，确认前置 release gates 全绿，唯一失败点是 `build-web-dist` 的 `Build web dists` 步骤；本地以 `env -u RUSTC_WRAPPER cargo build --target wasm32-unknown-unknown --manifest-path crates/oasis7_client_launcher/Cargo.toml --release --bin oasis7_client_launcher` 复现同一 `E0432 unresolved import oasis7::simulator::ProviderCompatibilityStatus` 签名。
+- 完成内容: 将 launcher 展示层使用的 `ProviderCompatibilityStatus` 收口为 `crates/oasis7_client_launcher/src/main.rs` 内部跨目标类型，并在 `crates/oasis7_client_launcher/src/launcher_core.rs` 的 native provider 检查结果里显式做 `into()` 映射，避免 wasm 目标再直接依赖 `oasis7::simulator` 的 native-only re-export。
+- 完成内容: 本地完成两层回归：`cargo build --target wasm32-unknown-unknown` 与 `cd crates/oasis7_client_launcher && env -u NO_COLOR trunk build --release --dist ../../output/release/web-launcher-dist` 均通过，确认 `build-web-dist` 的 launcher wasm 编译面已恢复。
+- 遗留事项: 仍需完成 `doc-governance-check` / `pm lint` / `git diff --check` / snapshot review / commit / PR / 合入后重新打 release tag。
+
+## 2026-04-13 23:44:45 CST / viewer_engineer
+- 完成内容: 已执行 `./scripts/doc-governance-check.sh`、`./scripts/pm/lint.sh`、`git diff --check` 与 `./scripts/pm/codex-review-snapshot.sh`；snapshot review 额外复跑 native/wasm `cargo check` 与 provider check tests，结论为未发现新增回归。
+- 完成内容: 已执行 `./scripts/pm/workflow-report.sh --phase close --role viewer_engineer --task-uid task_1d90a16ad56f4c4e96bfa5370199b3d7`，并将 task 状态从 `committed` 迁移到 `done`，使 `.pm` 真值、execution log 与 `doc/testing/project.md` 保持一致。
+- 遗留事项: 等待提交、发起 PR、合入后重新创建 GitHub release tag。
